Ripping the chat input box into another component. This commit only seperates the chat input into a new component.
